requirements.txt
streamlit==1.32.0
neo4j==5.18.0
openai==1.12.0
python-dotenv==1.0.1
plotly==5.19.0
pandas==2.2.1
scikit-learn==1.4.1
pyvis==0.3.2
streamlit-folium==0.18.0
folium==0.16.0
numpy==1.26.4
snowflake-connector-python==3.7.1
requests==2.31.0

Flask==2.3.3
flask-cors==4.0.0
Flask-JWT-Extended==4.5.2
SQLAlchemy==2.0.21
PyJWT==2.8.0
Werkzeug==2.3.7
python-dotenv==1.0.0
cryptography==41.0.4
pyotp==2.9.0
argon2-cffi==23.1.0

# Database drivers (install based on your database choice)
# PostgreSQL
psycopg2-binary==2.9.7

# MySQL
# pymysql==1.1.0

# Development and testing
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0
black==23.9.1
flake8==6.1.0

# Production server
gunicorn==21.2.0
